Added a retry decorator and an ADC read function.
This commit includes a minor improvement which consists of a retry decorator
that will attempt to get a sensor reading a given number of times, and throw
an exception if a reading is not read out within the retry period. This commit
also includes an added utility function to read the values from the MCP3008 ADC
chip through the SPI hardware interface on the Raspberry Pi.