Refactored toolchain.
This commit refactored the old airU.py script so that it acts as a toolchain of 3 support scripts. The data-dumpy.py script is used to collect data from the sensors and write the results to a local database. The data-upload.py script is used to upload data to the AirU central database used for the web front end. The data-dump.py script can be used by an active user to get data recently collected off of the device.